+++ EssentialFeed/EssentialFeed/FeedApi/FeedItemsMapper.swift
@@ -6,12 +6,7 @@
 //
 import Foundation
 
-internal struct RemoteFeedItem: Decodable {
-    internal let id: UUID
-    internal let description: String?
-    internal let location: String?
-    internal let image: URL
-}
+
 
 internal enum FeedItemsMapper {
