+++ EssentialFeed/EssentialFeed/FeedApi/RemoteFeedItem.swift
@@ -1,0 +1,14 @@
+//
+//  RemoteFeedItem.swift
+//  EssentialFeed
+//
+//  Created by Cristian Felipe Patiño Rojas on 3/4/25.
+//
+import Foundation
+
+internal struct RemoteFeedItem: Decodable {
+    internal let id: UUID
+    internal let description: String?
+    internal let location: String?
+    internal let image: URL
+}
